php session fixing.
the $_SESSION was working as usual on my wamp server. But, i just don't see it wasn't working on my internette server! 
so this update ams to fix this issue.
(now session just work find on wamp and on my server so evrebody can have the full exeperience whitout a need of a wamp server but with google crome!)